refactor(dialog-card): rename to index.vue, add float_header prop
Renaming to index.vue matches the directory-as-component convention
used elsewhere. float_header takes the header out of flow (absolutely
pinned to the top) so body content can center against the card's full
height instead of the space left over after the header — opt-in since
most existing callers want the header to push content down as usual.
