+++ src/components/modals/move-cards.vue
@@ -6,7 +6,7 @@ import { useI18n } from 'vue-i18n'
 import UiRadio from '@/components/ui-kit/radio.vue'
 import UiButton from '@/components/ui-kit/button.vue'
 import UiTappable from '@/components/ui-kit/tappable.vue'
-import DialogCard from '@/components/layout-kit/dialog-card/dialog-card.vue'
+import DialogCard from '@/components/layout-kit/dialog-card/index.vue'
 import GroupedList from '@/components/layout-kit/grouped-list.vue'
 import ScrollBar from '@/components/ui-kit/scroll-bar.vue'
 import { TYPE_SFX } from '@/sfx/config'
