+++ src/components/study-session/index.vue
@@ -3,7 +3,7 @@ import SessionFlashcard from './session-flashcard/index.vue'
 import SessionSummary from './session-summary/index.vue'
 import { computed, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
-import DialogCard from '@/components/layout-kit/dialog-card/dialog-card.vue'
+import DialogCard from '@/components/layout-kit/dialog-card/index.vue'
 import DialogCardPager from '@/components/layout-kit/dialog-card/dialog-card-pager.vue'
 import { emitSfx, emitStudySfx } from '@/sfx/bus'
 import { useProvideDeckContext } from './deck-context'
